+++ src/main/java/net/soulsweaponry/entity/effect/PostureBreak.java
@@ -1,10 +1,12 @@
 package net.soulsweaponry.entity.effect;
 
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
+import net.soulsweaponry.registry.EffectRegistry;
 
 public class PostureBreak extends StatusEffect {
 
@@ -21,11 +23,21 @@ public class PostureBreak extends StatusEffect {
             return true;
          }
     }
-    
+
     @Override
-    public void applyUpdateEffect(LivingEntity entity, int amplifier) {
-        entity.addStatusEffect(new StatusEffectInstance(StatusEffects.SLOWNESS, 20, 3));      
-        entity.addStatusEffect(new StatusEffectInstance(StatusEffects.WEAKNESS, 20, 9));      
-        entity.addStatusEffect(new StatusEffectInstance(StatusEffects.MINING_FATIGUE, 20, 9));      
+    public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
+        super.onApplied(entity, attributes, amplifier);
+        int duration = entity.hasStatusEffect(EffectRegistry.POSTURE_BREAK) ? entity.getStatusEffect(EffectRegistry.POSTURE_BREAK).getDuration() : 60;
+        entity.addStatusEffect(new StatusEffectInstance(StatusEffects.SLOWNESS, duration, 3));
+        entity.addStatusEffect(new StatusEffectInstance(StatusEffects.WEAKNESS, duration, 9));
+        entity.addStatusEffect(new StatusEffectInstance(StatusEffects.MINING_FATIGUE, duration, 9));
+    }
+
+    @Override
+    public void onRemoved(LivingEntity entity, AttributeContainer attributes, int amplifier) {
+        super.onRemoved(entity, attributes, amplifier);
+        if (entity.hasStatusEffect(StatusEffects.SLOWNESS)) entity.removeStatusEffect(StatusEffects.SLOWNESS);
+        if (entity.hasStatusEffect(StatusEffects.WEAKNESS)) entity.removeStatusEffect(StatusEffects.WEAKNESS);
+        if (entity.hasStatusEffect(StatusEffects.MINING_FATIGUE)) entity.removeStatusEffect(StatusEffects.MINING_FATIGUE);
     }
 }